+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.26] - 2022-11-29
+### Added
+* ListExtensions function to get the ToString of all elements in list
+### Changed
+* OrbitScript such that all of its fields are public
+
 ## [0.0.25] - 2022-11-16
 ### Added
 * ListExtensions to Extensions directory